feat(client): allow to add a closure as a middleware in client
